web-panel/requirements.txt
Flask>=3.0.0
Flask-Caching>=2.1.0
SQLAlchemy>=2.0.0
PyMySQL>=1.1.0
python-dotenv>=1.0.0
Jinja2>=3.1.0
gunicorn>=21.2.0
